test(ui): consolidate transcript video regressions
